add protocol control
