+++ memredis.go
@@ -7,18 +7,18 @@ import (
 )
 
 func Run () {
-	port := 3005
+	port := 3009
 	addr := syscall.SockaddrInet4{Port: port}
 	copy(addr.Addr[:], net.ParseIP("0.0.0.0").To4())
 	fmt.Println(addr.Addr[:])
-	peers := NewServerPeers("0.0.0.0:3005" , nil)
+	peers := NewServerPeers("0.0.0.0:3009" , nil)
 	//peers.SetPeers("0.0.0.0:3022", "0.0.0.0:3023")
 
-	commandHandler := NewCommandHandler(peers, Cachetable)
+	protocolHandler := NewProtocolHandler(peers, Cachetable)
 
 	//init worker pool
 	workerpool := NewWorkerPool(workerln)
-	workerpool.Run(commandHandler)
+	workerpool.Run(protocolHandler)
 
 	socket_server := &SocketServer{&addr}
 	socket_server.Serve()
